+++ utils/get_outstanding_paperwork.py
@@ -8,5 +8,5 @@ for emp in Employee.objects.filter(is_active=True):
             is_good = False
             paperwork_missing.append(paperwork.form.form_name)
     if not is_good:
-        # print(f'{emp.first_name} {emp.last_name} - {", ".join(paperwork_missing)}')
-        print(f"{emp.first_name} {emp.last_name}")
+        # print(f'{(emp.preferred_name if emp.preferred_name else emp.first_name)} {emp.last_name} - {", ".join(paperwork_missing)}')
+        print(f"{(emp.preferred_name if emp.preferred_name else emp.first_name)} {emp.last_name}")